Create README.md
This project is a data analysis for examining the performance in trial vs control stores to provide a recommendation for each location based on the  insight. of quantium stores. 
Select control stores – explore the data and define metrics for control store selection – think about what would make them a control store. Look at the drivers and make sure to  visualise these in a graph to better determine if they are suited. The tasks presented in using R code.